# docs/openapi.yaml
openapi: "3.0.0"
info:
  title: WhatsApp API MultiDevice
  version: 5.4.0
  description: This API is used for sending whatsapp via API
servers:
  - url: http://localhost:3000
...
security:
  - basicAuth: []
paths:
...
  /group:
    post:
      operationId: createGroup
      tags:
        - group
      summary: Create group and add participant
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: 'Example Group Title'
                participants:
                  type: array
                  items:
                    type: string
                    example:
                      - '6819241294719274'
                      - '6829241294719274'
                      - '6839241294719274'
                  example:
                    - '6819241294719274'
                    - '6829241294719274'
                    - '6839241294719274'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGroupResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/participants:
    post:
      operationId: addParticipantToGroup
      tags:
        - group
      summary: Adding more participants to group
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManageParticipantRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManageParticipantResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/participants/remove:
    post:
      operationId: removeParticipantFromGroup
      tags:
        - group
      summary: Remove participants from group
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManageParticipantRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManageParticipantResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/participants/promote:
    post:
      operationId: promoteParticipantToAdmin
      tags:
        - group
      summary: Promote participants to admin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManageParticipantRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManageParticipantResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/participants/demote:
    post:
      operationId: demoteParticipantToMember
      tags:
        - group
      summary: Demote participants to member
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManageParticipantRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManageParticipantResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/join-with-link:
    post:
      operationId: joinGroupWithLink
      tags:
        - group
      summary: Join group with link
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                link:
                  type: string
                  example: 'https://chat.whatsapp.com/whatsappKeyJoinGroup'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/participant-requests:
    get:
      operationId: getGroupParticipantRequests
      tags:
        - group
      summary: Get list of participant requests to join group
      parameters:
        - name: group_id
          in: query
          required: true
          schema:
            type: string
          example: '120363024512399999@g.us'
          description: The group ID to get participant requests for
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupParticipantRequestListResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/participant-requests/approve:
    post:
      operationId: approveGroupParticipantRequest
      tags:
        - group
      summary: Approve participant request to join group
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                group_id:
                  type: string
                  example: '120363024512399999@g.us'
                  description: The group ID
                participant_id:
                  type: string
                  example: '6281234567890'
                  description: The participant's WhatsApp ID to approve
              required:
                - group_id
                - participant_id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/participant-requests/reject:
    post:
      operationId: rejectGroupParticipantRequest
      tags:
        - group
      summary: Reject participant request to join group
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                group_id:
                  type: string
                  example: '120363024512399999@g.us'
                  description: The group ID
                participant_id:
                  type: string
                  example: '6281234567890'
                  description: The participant's WhatsApp ID to reject
              required:
                - group_id
                - participant_id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
  /group/leave:
    post:
      operationId: leaveGroup
      tags:
        - group
      summary: Leave group
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                group_id:
                  type: string
                  example: '120363024512399999@g.us'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBadRequest'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServer'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    CreateGroupResponse:
      type: object
      properties:
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: Success get list groups
        results:
          type: object
          properties:
            group_id:
              type: string
              example: 1203632782168851111@g.us
    ManageParticipantRequest:
      type: object
      properties:
        group_id:
          type: string
          example: 1203632782168851111@g.us
        participants:
          type: array
          items:
            type: string
          example:
            - '6819241294719274'
            - '6829241294719274'
            - '6839241294719274'
    ManageParticipantResponse:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: Success get list groups
        results:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              participant:
                type: string
                example: '6289987391723@s.whatsapp.net'
              status:
                type: string
                example: success
              message:
                type: string
                example: Participant added
...
    GenericResponse:
      type: object
      properties:
        code:
          type: string
          example: SUCCESS
        message:
          type: string
          example: Success
        results:
          type: string
          example: null
    ErrorInternalServer:
      type: object
      properties:
        code:
          type: string
          example: INTERNAL_SERVER_ERROR
          description: 'SYSTEM_CODE_ERROR'
        message:
          type: string
          example: you are not loggin
          description: 'Detail error message'
        results:
          type: object
          example: null
          description: 'additional data'
    ErrorBadRequest:
      type: object
      properties:
        code:
          type: string
          example: 400
          description: 'HTTP Status Code'
        message:
          type: string
          example: field cannot be blank
          description: 'Detail error message'
        results:
          type: object
          example: null
          description: 'additional data'
...
    GroupParticipantRequestListResponse:
      type: object
      properties:
        code:
          type: string
          example: "SUCCESS"
        message:
          type: string
          example: "Success getting list requested participants"
        results:
          type: object
          properties:
            data:
              type: array
              items:
                type: object
                properties:
                  jid:
                    type: string
                    example: "6289685024091@s.whatsapp.net"
                  requested_at:
                    type: string
                    format: date-time
                    example: "2024-10-11T21:27:29+07:00"